requirements.txt
argparse
pyyaml
regex
